requirements.txt
albumentations==1.4.2
easyocr==1.7.1
hydralit-components==1.0.10
Markdown==3.6
numpy==1.26.4
opencv-python==4.9.0.80
Pillow==9.3.0
protobuf==3.20.1
stqdm==0.0.5
streamlit==1.32.2
streamlit-lottie==0.0.5
tqdm==4.66.2
ultralytics==8.1.36